requirements.txt
# Core dependencies
ccxt>=4.0.0
aiohttp>=3.8.0
pydantic>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Technical analysis
ta>=0.10.0

# Configuration
pyyaml>=6.0

# Optional: For enhanced logging
python-dotenv>=1.0.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0
